+++ core/interfaces/repository/IBaseRepository.cs
@@ -4,9 +4,9 @@ namespace Gestion.core.interfaces.repository;
 
 public interface IBaseRepository<T> where T : IModel
 {
-    Task<T?> FindById(int id);
+    Task<T?> FindById(long id);
     Task<List<T>> FindAll();
-    Task<bool> DeleteById(int id);
+    Task<bool> DeleteById(long id);
     Task<bool> Update(T entity); // Actualiza
     Task<bool> Save(T entity);   // Crea
 }
