+++ core/interfaces/service/IBaseService.cs
@@ -2,9 +2,9 @@ namespace Gestion.core.interfaces.service;
 
 public interface IBaseService<T>
 {
-    Task<T?> FindById(int id);
+    Task<T?> FindById(long id);
     Task<List<T>> FindAll();
-    Task<bool> DeleteById(int id);
+    Task<bool> DeleteById(long id);
     Task<bool> Update(T entity);
     Task<bool> Save(T entity);
 }
