+++ core/model/Usuario.cs
@@ -4,11 +4,11 @@ namespace Gestion.core.model;
 
 public class Usuario : IModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string NombreUsuario { get; private set; } = string.Empty;
     public string Contraseña { get; private set; } = string.Empty;
 
-    public Usuario(int id, string nombreUsuario, string contraseña)
+    public Usuario(long id, string nombreUsuario, string contraseña)
     {
         this.Id = Id;
         this.NombreUsuario = nombreUsuario;
